+++ Intersect (Core)/GameObjects/NpcBase.cs
@@ -232,7 +232,7 @@ namespace Intersect.GameObjects
 
         public Guid ItemId;
 
-        public int Quantity;
+        public long Quantity;
 
     }
