+++ Intersect (Core)/Network/Packets/Server/BagUpdatePacket.cs
@@ -6,7 +6,7 @@ namespace Intersect.Network.Packets.Server
     public class BagUpdatePacket : InventoryUpdatePacket
     {
 
-        public BagUpdatePacket(int slot, Guid id, int quantity, Guid? bagId, int[] statBuffs) : base(
+        public BagUpdatePacket(int slot, Guid id, long quantity, Guid? bagId, int[] statBuffs) : base(
             slot, id, quantity, bagId, statBuffs
         )
         {
